Add Limit for Input & Fix Anything Run But Have Many Bugs

+++ script.js
@@ -33,6 +33,7 @@ themeBtn.addEventListener("click", toggleTheme);
 // Calculator Elements Select & Variable
 const displayInput = document.getElementById("display-input");
 const buttons = document.querySelectorAll("button.row");
+const validInput = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9","+","-","*","/"];
 let operatorAddToInput = { "+": 0, "-": 0, "*": 0, "/": 0 };
 const operatorKeys = Object.keys(operatorAddToInput);
 let result = "";
@@ -54,13 +55,27 @@ const buttonAction = (e) => {
       return;
 
     case targetText === "=":
-      result = displayInput.value = eval(displayInput.value);
-      displayInput.value = result;
-      localStorage.setItem("lastResult", result);
+      operatorKeys.forEach((operator) => (operatorAddToInput[operator] = 0));
+      
+      function checkEntry(entry){
+        const entryArray = entry.split("")
+        const entryExisted = entryArray.map((entryLetter)=>validInput.includes(entryLetter));
+
+        if (entryExisted.includes(false)) {
+          console.log("entryExisted Have False Item");
+          return;
+        } else {
+          console.log(entry);
+          result = eval(entry);
+          displayInput.value = result;
+          localStorage.setItem("lastResult", result);
+        }
+      }
+      checkEntry(displayInput.value)
       break;
 
     case targetText === "AC":
-      operatorKeys.forEach( (operator) => operatorAddToInput[operator]=0 )
+      operatorKeys.forEach((operator) => (operatorAddToInput[operator] = 0));
       result = "";
       displayInput.value = result;
 
@@ -69,7 +84,14 @@ const buttonAction = (e) => {
 
     case targetText === "DEL":
       const lastChar = result.toString().slice(result.length - 1);
-      if(lastChar === "+" || lastChar === "-" || lastChar === "*" || lastChar === "/"){operatorAddToInput[lastChar]--}
+      if (
+        lastChar === "+" ||
+        lastChar === "-" ||
+        lastChar === "*" ||
+        lastChar === "/"
+      ) {
+        operatorAddToInput[lastChar]--;
+      }
       result = result.toString().slice(0, -1);
       displayInput.value = result;
       localStorage.setItem("lastResult", result);
@@ -86,8 +108,13 @@ const buttonAction = (e) => {
       localStorage.setItem("lastResult", result);
       break;
 
-    case targetText === "+" || targetText === "-" || targetText === "*" || targetText === "/":
-      if(result ===""){return}
+    case targetText === "+" ||
+      targetText === "-" ||
+      targetText === "*" ||
+      targetText === "/":
+      if (result === "") {
+        return;
+      }
       for (let i = 0; i < operatorKeys.length; i++) {
         if (operatorAddToInput[operatorKeys[i]]) {
           return;
